refactor: merge `extractComments` option tests
